+++ tigerpm/fft.hpp
@@ -15,9 +15,9 @@
 void fft3d_init(int N);
 void fft3d_execute();
 void fft3d_inv_execute();
-vector<float> fft3d_read_real(const range<int>&);
+vector<double> fft3d_read_real(const range<int>&);
 vector<cmplx> fft3d_read_complex(const range<int>&);
-void fft3d_accumulate_real(const range<int>&, const vector<float>&);
+void fft3d_accumulate_real(const range<int>&, const vector<double>&);
 void fft3d_accumulate_complex(const range<int>&, const vector<cmplx>&);
 void fft3d_destroy();
 void fft3d_force_real();
